+++ CleverPush/Source/CPAppBannerModuleInstance.h
@@ -35,7 +35,6 @@
 - (void)triggerEvent:(NSString *)eventId properties:(NSDictionary *)properties;
 - (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId;
 - (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId force:(BOOL)force;
-- (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId notificationId:(NSString*)notificationId;
 - (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId notificationId:(NSString*)notificationId force:(BOOL)force;
 - (void)startup;
 - (NSMutableArray*)shownAppBanners;
@@ -53,7 +52,6 @@
 - (void)scheduleBannersForEvent:(NSString *)eventId fromActiveBanners:(NSArray<CPAppBanner *> *)activeBanners;
 - (void)scheduleBannersForNoEventFromActiveBanners:(NSArray<CPAppBanner *> *)activeBanners;
 - (NSTimeInterval)calculateDelayForBanner:(CPAppBanner *)banner;
-- (void)showBanner:(CPAppBanner*)banner;
 - (void)presentAppBanner:(CPAppBannerViewController*)appBannerViewController  banner:(CPAppBanner*)banner;
 - (void)showNextActivePendingBanner:(CPAppBanner*)banner;
 - (void)sendBannerEvent:(NSString*)event forBanner:(CPAppBanner*)banner forScreen:(CPAppBannerCarouselBlock*)screen forButtonBlock:(CPAppBannerButtonBlock*)button forImageBlock:(CPAppBannerImageBlock*)image blockType:(NSString*)type;
@@ -92,7 +90,6 @@
 + (void)setBannersForDeepLink:(NSMutableArray*)appBanner;
 + (NSMutableArray*)getBannersForDeepLink;
 + (void)updateBannersForDeepLinkWithURL:(NSURL*)url;
-+ (void)setSilentPushAppBannersIds:(NSString*)appBannerId notificationId:(NSString*)notificationId;
-+ (NSMutableArray*)getSilentPushAppBannersIds;
++ (void)addSilentPushAppBannersId:(NSString*)appBannerId notificationId:(NSString*)notificationId;
 
 @end