+++ Frameworks/CleverPush.xcframework/ios-arm64_x86_64-maccatalyst/CleverPush.framework/Versions/A/Headers/CPUtils.h
@@ -47,7 +47,6 @@
 + (NSURL *)removeQueryParametersFromURL:(NSURL *)url;
 + (NSDictionary *)convertConnectionOptionsToLaunchOptions:(UISceneConnectionOptions* )connectionOptions  API_AVAILABLE(ios(13.0));
 + (UIImage *)resizedImageNamed:(NSString *)imageName withSize:(CGSize)newSize;
-+ (NSCache *)sharedImageCache;
 
 /**
  * Returns a shared NSCache instance for caching images across the app